restructured io folders and make sure binaries are first in both nbody6 and nbody6pp
