+++ clustertools/cluster/__init__.py
@@ -10,6 +10,8 @@ to_nbody=operations.to_nbody
 to_radec=operations.to_radec
 to_galpy=operations.to_galpy
 to_units=operations.to_units
+to_sudays=operations.to_sudays
+to_audays=operations.to_audays
 to_centre=operations.to_centre
 to_cluster=operations.to_cluster
 to_galaxy=operations.to_galaxy
