+++ tests/test_orbits.py
@@ -1,0 +1,28 @@
+import clusertools as ctools
+
+def test_initialize_orbit(tol=0.1):
+	pass
+
+def test_initialize_orbits(tol=0.1):
+	pass
+
+def test_integrate_orbit(tol=0.1):
+	pass
+
+def test_integrate_orbits(tol=0.1):
+	pass
+
+def test_orbit_interpolate(tol=0.1):
+	pass
+
+def test_orbital_path(tol=0.1):
+	pass
+
+def test_orbital_path_match(tol=0.1):
+	pass
+
+def test_calc_actions(tol=0.1):
+	pass
+	
+def test_ttensor(tol=0.1):
+	pass
